Experimental

+++ Class/cli.py
@@ -3,7 +3,7 @@ from Class.wireguard import Wireguard
 from Class.templator import Templator
 from Class.base import Base
 from Class.bird import Bird
-import subprocess, logging, sys, os
+import subprocess, logging, time, sys, os
 
 class CLI(Base):
 
@@ -23,9 +23,18 @@ class CLI(Base):
         self.wg = Wireguard(self.path)
         self.wg.bender()
 
-    def connect(self,dest,token,linkType="default",port=51820):
+    def connect(self,dest,token,linkType="default",port=51820,network=""):
         self.wg = Wireguard(self.path)
-        self.wg.connect(dest,token,linkType,port)
+        status = self.wg.connect(dest,token,linkType,port,network)
+        if self.wg.getInitial:
+            if not status['v4'] and not status['v6']:
+                print(f"Initial link wasn't setup.")
+                return 
+            print("Waiting for meshing to complete.")
+            for i in range(1, 300):
+                if os.path.isfile(f"{self.path}/configs/state.json"): return
+                time.sleep(1)
+            print("Meshing seems to have failed.")
 
     def proximity(self,cutoff=0):
         self.wg = Wireguard(self.path)
@@ -45,13 +54,27 @@ class CLI(Base):
     def update(self):
         subprocess.run("cd; git pull",shell=True)
 
-    def clean(self):
+    def clean(self,ignoreJSON,ignoreEndpoint):
         self.wg = Wireguard(self.path)
-        self.wg.clean()
+        self.wg.clean(ignoreJSON,ignoreEndpoint)
 
     def migrate(self):
-        self.wg = Wireguard(self.path)
+        self.wg = Wireguard(self.path,False,True)
         self.wg.updateConfig()
+
+    def geo(self):
+        headers = {"Origin":"https://ip-api.com"}
+        geoDataRaw = self.call("https://demo.ip-api.com/json/?fields=66842623",{},"GET",headers)
+        if geoDataRaw:
+            geoData = geoDataRaw.json()
+            config = self.readJson(f'{self.path}/configs/config.json')
+            if not "geo" in config: config['geo'] = {}
+            config['geo']['countryCode'] = geoData['countryCode']
+            config['geo']['continent'] = geoData['continent']
+            config['geo']['country'] = geoData['country']
+            config['geo']['city'] = geoData['city']
+            print(f"Updated geodata {config['geo']}")
+            self.saveJson(config,f"{self.path}/configs/config.json")
 
     def recover(self):
         stream_handler = logging.StreamHandler()
@@ -62,11 +85,12 @@ class CLI(Base):
         self.bird.bird(True)
 
     def token(self):
-        if os.path.isfile(f"{self.path}/token"):
-            with open(f'{self.path}/token') as f:
-                print(f.read().rstrip())
+        tokens = self.readJson(f"{self.path}/tokens.json")
+        if tokens:
+            print(f"Connect: {', '.join(tokens['connect'])}")
+            print(f"Peer: {', '.join(tokens['peer'])}")
         else:
-            print("Unable to load the token file")
+            print("Unable to load the tokens.json")
 
     def status(self):
         print("--- Services ----")
@@ -78,9 +102,12 @@ class CLI(Base):
         wgmeshBirdRunning = "wgmesh-bird is not running." if not "running" in proc else "wgmesh-bird is running."
         print(f"{birdRunning}\t{wgmeshRunning}\t{wgmeshBirdRunning}")
         print("--- Wireguard ---")
-        network = self.readConfig(f"{self.path}/configs/network.json")
-        print("Destination\tPacketloss\tJitter")
-        jittar,loss = 0,0
+        network = self.readJson(f"{self.path}/configs/network.json")
+        if not network:
+            print("Unable to load network.json")
+            return
+        print("Destination\tStatus\tPacketloss\tJitter")
+        jittar,loss,online,offline = 0,0,0,0
         for dest,data in network.items():
             hasLoss,hasJitter = "No","No"
             if dest == "updated": continue
@@ -90,12 +117,20 @@ class CLI(Base):
             if data['jitter']:
                 hasJitter = "Yes"
                 jittar += 1
-            print(f"{dest}\t{hasLoss}\t\t{hasJitter}")
-        print(f"{len(network) -1}\t\t{loss}\t\t{jittar}")
-
+            if data['state']:
+                state = "Online"
+                online += 1
+            else:
+                state = "Offline"
+                offline += 1
+            print(f"{dest}\t{state}\t{hasLoss}\t\t{hasJitter}")
+        print(f"{len(network) -1}\t\t{online}/{offline}\t{loss}\t\t{jittar}")
 
     def disable(self,option):
-        config = self.readConfig(f"{self.path}/configs/config.json")
+        config = self.readJson(f"{self.path}/configs/config.json")
+        if not config:
+            print("Unable to load config.json")
+            return
         if "mesh" in option:
             self.wg.saveJson({},f"{self.path}/configs/state.json")
         elif "ospfv2" in option:
@@ -104,18 +139,28 @@ class CLI(Base):
             config['bird']['ospfv3'] = False
         elif "client" in option:
             config['bird']['client'] = False
+        elif "notifications" in option:
+            config['notifications']['enabled'] = False
         elif "wgobfs" in option:
             if "wgobfs" in config['linkTypes']: config['linkTypes'].remove("wgobfs")
         elif "ipt_xor" in option:
             if "ipt_xor" in config['linkTypes']: config['linkTypes'].remove("ipt_xor")
+        elif "amneziawg" in option:
+            if "amneziawg" in config['linkTypes']: config['linkTypes'].remove("amneziawg")
         else:
-            print("Valid options: mesh, ospfv2, ospfv3, wgobfs, ipt_xor, client")
+            print("Valid options: mesh, ospfv2, ospfv3, wgobfs, ipt_xor, amneziawg, client, notifications")
             return
-        self.saveJson(config,f"{self.path}/configs/config.json")
+        response = self.saveJson(config,f"{self.path}/configs/config.json")
+        if not response:
+            print("Failed to save config.json")
+            return
         print("You should reload the services to apply any config changes")
             
     def enable(self,option):
-        config = self.readConfig(f"{self.path}/configs/config.json")
+        config = self.readJson(f"{self.path}/configs/config.json")
+        if not config:
+            print("Unable to load config.json")
+            return
         if "mesh" in option:
             if os.path.isfile(f"{self.path}/configs/state.json"): os.remove(f"{self.path}/configs/state.json")
         elif "ospfv2" in option:
@@ -124,38 +169,61 @@ class CLI(Base):
             config['bird']['ospfv3'] = True
         elif "client" in option:
             config['bird']['client'] = True
+        elif "notifications" in option:
+            config['notifications']['enabled'] = True
         elif "wgobfs" in option:
             if not "wgobfs" in config['linkTypes']: config['linkTypes'].append("wgobfs")
             print("You still need to install wgobfs with: bash /opt/wg-mesh/tools/wgobfs.sh")
         elif "ipt_xor" in option:
             if not "ipt_xor" in config['linkTypes']: config['linkTypes'].append("ipt_xor")
             print("You still need to install ipt_xor with: bash /opt/wg-mesh/tools/xor.sh")
+        elif "amneziawg" in option:
+            if not "amneziawg" in config['linkTypes']: config['linkTypes'].append("amneziawg")
+            print("You still need to install amneziawg with: bash /opt/wg-mesh/tools/amnezia.sh")
         else:
-            print("Valid options: mesh, ospfv2, ospfv3, wgobfs, ipt_xor, client")
+            print("Valid options: mesh, ospfv2, ospfv3, wgobfs, ipt_xor, amneziawg, client, notifications")
             return
-        self.saveJson(config,f"{self.path}/configs/config.json")
+        response = self.saveJson(config,f"{self.path}/configs/config.json")
+        if not response:
+            print("Failed to save config.json")
+            return
         print("You should reload the services to apply any config changes")
 
     def setOption(self,options):
-        validOptions = ["area","prefix","defaultLinkType","basePort","tick","subnet"]
+        validOptions = ["area","prefix","defaultLinkType","basePort","tick","reloadInterval","reloadPercentage","operationMode","vxlanOffset","subnet","subnetVXLAN","subnetLinkLocal","AllowedPeers","gotifyUp","gotifyDown","gotifyError",'gotifyDiag']
         if len(sys.argv) == 0:
             print(f"Valid options: {', '.join(validOptions)}")
         else:
             key, value = options
             if key in validOptions:
-                config = self.readConfig(f"{self.path}/configs/config.json")
-                if key == "basePort":
+                config = self.readJson(f"{self.path}/configs/config.json")
+                if not config:
+                    print(f"Unable to read config.json")
+                    return
+                if key == "basePort" or key == "vxlanOffset" or key == "operationMode":
                     config[key] = int(value)
-                elif key == "area" or key == "tick":
+                elif key == "area" or key == "tick" or key == "reloadInterval" or key == "reloadPercentage":
                     config['bird'][key] = int(value)
+                elif key == "gotifyUp" or key == "gotifyDown" or key == "gotifyError" or key == "gotifyDiag":
+                    config['notifications'][key] = value
+                elif key == "AllowedPeers":
+                    if value in config['AllowedPeers']:
+                        config['AllowedPeers'].remove(value)
+                    else:
+                        config['AllowedPeers'].append(value)
                 else:
                     config[key] = value
-                self.saveJson(config,f"{self.path}/configs/config.json")
+                response = self.saveJson(config,f"{self.path}/configs/config.json")
+                if not response:
+                    print("Failed to save config.json")
+                    return
                 print("You should reload the services to apply any config changes")
-                if key == "subnet":        
+                if key == "subnet" or key == "subnetVXLAN":        
                     print("Reconfiguring dummy")
                     self.wg = Wireguard(self.path)
                     self.wg.reconfigureDummy()
             else:
                 print(f"Valid options: {', '.join(validOptions)}")
         
+    def cost(self,link,cost=0):
+        self.wg.setCost(link,cost)
